icu-system/backend-api/requirements.txt
# FastAPI backend API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Kafka
confluent-kafka==2.3.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0